+++ backend/src/domain/accounting/__init__.py
@@ -1,0 +1,20 @@
+"""Domain accounting package."""
+
+from src.domain.accounting.transaction_type import TransactionType
+from src.domain.accounting.tax_lot import TaxLot, LotDepletion, LotStatus
+from src.domain.accounting.transaction import Transaction
+from src.domain.accounting.fifo_engine import FIFOMatcher, InsufficientHoldingsError
+from src.domain.accounting.portfolio_replayer import PortfolioReplayer, HoldingSnapshot, PortfolioValuation
+
+__all__ = [
+    "TransactionType",
+    "TaxLot",
+    "LotDepletion",
+    "LotStatus",
+    "Transaction",
+    "FIFOMatcher",
+    "InsufficientHoldingsError",
+    "PortfolioReplayer",
+    "HoldingSnapshot",
+    "PortfolioValuation",
+]